Use micromatcher for pattern matching
Micromatcher supports better glob patterns for matching files
even if the user shell does not. Previous implementation relied on grep.

Now, we ask Git to give us the changed files files only once with execa.

Future-proofed the matching implementation with micromatch.some()
for matching multiple patterns for a single set of commands
when we upgrade the configuration to support it.

We no longer rely on bash scripts to execute multiple commands
simplifying and unifying the implementation.